+++ src/components/films-board.js
@@ -1,23 +1,8 @@
-import {createElement} from "../utilities";
-
 // контейнер для списка и секций Top rated & Most commented
-export default class FilmsBoard {
-  constructor() {
-    this._element = null;
-  }
+import AbstractComponent from "./abstract-component";
 
+export default class FilmsBoard extends AbstractComponent {
   getTemplate() {
     return `<section class="films"></section>`;
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
