+++ src/components/films-list.js
@@ -1,4 +1,4 @@
-import {createElement} from "../utilities";
+import AbstractComponent from "./abstract-component";
 
 const createFilmsListTemplate = () => {
   return (
@@ -9,24 +9,13 @@ const createFilmsListTemplate = () => {
   );
 };
 
-export default class FilmsList {
+export default class FilmsList extends AbstractComponent {
   constructor(film) {
+    super();
     this._film = film;
-    this._element = null;
   }
 
   getTemplate() {
     return createFilmsListTemplate(this._film);
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
