+++ src/components/no-films.js
@@ -1,22 +1,7 @@
-import {createElement} from "../utilities";
+import AbstractComponent from "./abstract-component";
 
-export default class NoFilms {
-  constructor() {
-    this._element = null;
-  }
-
+export default class NoFilms extends AbstractComponent {
   getTemplate() {
     return `<h2 class="films-list__title">There are no movies in our database</h2>`;
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
